Update project structure and enhance functionality
The project file has been updated with packing and publishing details. Adjustments to the build properties include changes in the target framework and enabling of implicit usings. Significant changes in MimeTypes involve adding a TryGetMimeTypeFromFile method and switching MimeType to a regular public record. A new README.md file was also added to provide installation and usage details of the library.